Unused and redundant code removed

+++ app/src/main/java/com/example/notepad/repository/NoteRepository.kt
@@ -2,11 +2,10 @@ package com.example.notepad.repository
 
 import android.content.Context
 import android.util.Log
-import android.widget.Toast
-import com.example.notepad.db.NoteDao
 import com.example.notepad.db.NoteDatabase
 import com.example.notepad.db.NoteModel
 
+// Interacts with the Room database
 class NoteRepository {
     suspend fun addNote(context: Context, title: String?, content: String?) {
         val newNote = NoteModel(title, content)
@@ -18,12 +17,14 @@ class NoteRepository {
     }
 
     suspend fun updateNote(context: Context, id: Int, title: String?, content: String?) {
-        var updatedNote = NoteModel(title, content)
+        val updatedNote = NoteModel(title, content)
         updatedNote.id = id
-        NoteDatabase(context).getNotesDao().updateNote(updatedNote!!)
+        NoteDatabase(context).getNotesDao().updateNote(updatedNote)
     }
 
-    suspend fun deleteNote(context: Context, id: Int, title: String?, content: String?) {
+    suspend fun deleteNote(context: Context, id: Int) {
+        Log.d("USER_TEST","Entering Delete method with ID: $id")
+        if(id==-1) return
         val deleteNote = NoteDatabase(context).getNotesDao().getNoteById(id)
         NoteDatabase(context).getNotesDao().deleteNote(deleteNote!!)
     }
